refactor:  improve type casting across various modules
- Updated memory allocation and casting in `alloc.h`, `boxed.h`, `list.h`, and `string.cc` to use `ptr::cast` and `mem::move` for better type safety and performance.
- Enhanced `Clap` parser to handle command line arguments more efficiently by using `Slice` for argument collection.
- Modified `HashMap` and `HashSet` to utilize `mem::move` for value transfers, improving efficiency.
- Introduced `FileName` struct in `path.cc` to encapsulate file name operations, simplifying extension and stem retrieval.
- Improved file and IO operations in `fs.inl` and `io.inl` to handle read/write operations with better error handling and type casting.
- Updated synchronization primitives in `sync` module to ensure proper memory management and type safety.
- Refined thread handling in `join.cc` to use safer pointer casting.
- General improvements in type casting and memory management across various files to enhance code clarity and performance.

+++ src/sfc/collections/hash/hash_map.cxx
@@ -48,14 +48,14 @@ SFC_TEST(map_try_insert) {
 }
 
 SFC_TEST(map_insert) {
-  auto t = HashMap<int, int>{};
-  for (auto i = 0; i < 10; ++i) {
+  auto t = HashMap<u32, u32>{};
+  for (auto i = 0U; i < 10U; ++i) {
     t.insert(i, i * 10);
     sfc::expect_eq(*t.get(i), i * 10);
-    sfc::expect_eq(t.len(), static_cast<usize>(i + 1));
+    sfc::expect_eq(t.len(), i + 1);
     t.insert(i, i * 100);
     sfc::expect_eq(*t.get(i), i * 100);
-    sfc::expect_eq(t.len(), static_cast<usize>(i + 1));
+    sfc::expect_eq(t.len(), i + 1);
   }
 
   sfc::expect_eq(t.len(), 10U);
